backend/requirements.txt
asgiref==3.2.3
attrs==19.3.0
Django==3.0.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
entrypoints==0.3
flake8==3.7.9
importlib-metadata==1.3.0
mccabe==0.6.1
more-itertools==8.0.2
packaging==19.2
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pyparsing==2.4.5
pytest==5.3.2
pytest-django==3.7.0
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
wcwidth==0.1.7
zipp==0.6.0
